backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic>=2.0.0
email-validator>=2.0.0
requests==2.31.0
python-dotenv==1.0.0
gitpython==3.1.40
bandit==1.7.6
openai>=2.8.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
authlib==1.3.0
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.9  # PostgreSQL driver (optional, only needed for PostgreSQL)
pymysql>=1.1.0  # MySQL driver (optional, only needed for MySQL)
httpx==0.25.2
stripe==7.0.0
reportlab>=4.0.7
Pillow>=10.1.0

